+++ R/synapse_version.R
@@ -9,6 +9,10 @@
 #' Specifies the version numbers available for each cancer cohort. Version
 #' numbers are specified as part of the call to `pull_data_synapse()`.
 #'
+#' @param cohort Vector specifying the cohort(s) of interest. Cohorts must be
+#'   one of "NSCLC" (Non-Small Cell Lung Cancer), "CRC" (Colorectal Cancer), or
+#'   "BrCa" (Breast Cancer), "PANC" (Pancreatic Cancer), "Prostate" (Prostate Cancer),
+#'   and "BLADDER" (Bladder Cancer).
 #' @param most_recent Indicates whether the function will return only the most
 #'   recent version number for each cohort (`most_recent` = TRUE) or all
 #'   available version numbers for each cohort (`most_recent`= FALSE)
@@ -22,9 +26,17 @@
 #'
 #' dplyr
 
-synapse_version <- function(most_recent = FALSE) {
+synapse_version <- function(cohort = NULL, most_recent = FALSE) {
   if (!is.logical(most_recent)) {
     stop("Please provide TRUE or FALSE for the most_recent argument")
+  }
+
+  # if cohort is not specified, return all cohorts
+  if (is.null(cohort)){
+    select_cohort <- c("NSCLC", "CRC", "BrCa", "BLADDER", "PANC", "Prostate")
+  } else {
+    select_cohort <- rlang::arg_match(cohort, c("NSCLC", "CRC", "BrCa", "BLADDER", "PANC", "Prostate"),
+                                    multiple = TRUE)
   }
 
   # define numeric release date
@@ -40,13 +52,15 @@ synapse_version <- function(most_recent = FALSE) {
   if (most_recent == FALSE) {
     synapse_tables_dts %>%
       group_by(.data$cohort, .data$version) %>%
-      filter(row_number() == 1) %>%
+      filter(.data$cohort %in% c(select_cohort),
+             row_number() == 1) %>%
       ungroup() %>%
       arrange(.data$cohort, .data$numeric_release_date) %>%
       select("cohort", "version", "release_date") %>%
       mutate(versions_returned = "All Versions")
   } else {
     synapse_tables_dts %>%
+      filter(.data$cohort %in% c(select_cohort)) %>%
       group_by(.data$cohort, .data$pubcon) %>%
       slice(which.max(.data$numeric_release_date)) %>%
       ungroup() %>%